backend/foodgram/requirements.txt
django==4.1.4
djangorestframework==3.14.0
Pillow==9.4.0
pytest==6.2.5
pytest_django==4.5.2
pytest_pythonpath==0.7.4
django-filter==22.1
psycopg2-binary==2.9.5
python-dotenv==0.21.0
django-cors-headers==3.13.0
gunicorn==20.1.0